SVIT
Satellite Vegetation Indices Trend (SVIT) is an Algorithm created by the PhD student Baldo Marco to analyze the forest changes in terms of cover and productivity. Together with PhD Piero Zannini, PhD student Kurian Ayushi and the Full Professor of the University of Bologna Duccio Rocchini, we create an R script to analyze the Dry Matter Productivity, Fraction of Absorbed Photosynthetically Active Radiation and Normalized Different Vegetation Index trends of a selected area, in this case over the Western Ghats world Biodiversity Hotspot under the theory of the multispectral bands' analysis and multiple time series analysis.